[PDI-11768] - Log distorted after overwriting the job using "Save As" in spoon 5.0.x
CAUSE:
the issue was caused by a lookup in LogTableFields collection while
the elements of the collection did not have overriden equals() and hashCode()
it worked until the job/trans has been "Saved As" as the object ref comparison worked.
After "Save as" lookup for another object failed.

RESOLUTION:
define equals() and hashCode() for LogTableField so that collection lookup is successful
two instances are equal if their ids are equals with assumption that there can't
be more than one field with the same id for a single log table